[numpy] add array access for dim = 1

+++ pythran/pythonic++/core/ndarray.h
@@ -36,9 +36,9 @@ namespace  pythonic {
                 impl::shared_ref< raw_array<T> > data;
                 long shape[N];
 
-                ndarray(std::initializer_list<int> s)
+                ndarray(std::initializer_list<size_t> s)
                 {
-                    long r = 1;
+                    size_t r = 1;
                     long * is = shape;
                     for(auto v :s )
                         r*=(*is++=v);
@@ -73,12 +73,12 @@ namespace  pythonic {
                         return offset(t0 * shape[N - sizeof...(Types) - 1] + t1, tn...); 
                     }
 
-                int offset(int t0, int t1)
+                long unsigned int offset(int t0, int t1)
                 {
                     return t0 * shape[N-1] + t1;
                 }
 
-                T* at(T* from, int t)
+                T* at(T* from, long unsigned int t)
                 {
                     return from +t;
                 }
